+++ packages/components-js/tests/unit/components/marque.spec.ts
@@ -3,10 +3,6 @@ import { Marque } from '../../../src/components/basic/marque/marque';
 
 describe('Component <p-marque>', () => {
 
-  it('should build', () => {
-    expect(new Marque()).toBeTruthy();
-  });
-
   it('should render correctly in default mode with shadow dom', async () => {
     const page = await newSpecPage({
       components: [Marque],
@@ -15,6 +11,5 @@ describe('Component <p-marque>', () => {
     expect(page.root.shadowRoot).toBeTruthy();
     expect(page.root.querySelector('.p-marque')).toBeFalsy();
     expect(page.root.shadowRoot.querySelector('.p-marque')).toBeTruthy();
-    expect(page.root).toMatchSnapshot();
   });
 });