+++ packages/components-js/src/components/form/checkbox-wrapper/checkbox-wrapper.tsx
@@ -38,7 +38,7 @@ export class CheckboxWrapper {
   private input: HTMLInputElement;
   private labelId = randomString();
 
-  public componentDidLoad() {
+  public componentWillLoad() {
     this.setInput();
     this.setState();
     this.bindStateListener();