+++ packages/storefront/src/components/Playground.vue
@@ -70,7 +70,7 @@
     @Prop({default: () => ({height: 'auto', spacing: 'none'})}) public childElementLayout!: ChildElementLayout;
 
     public theme: Theme = 'light';
-    public markup: string = '';
+    public markup = '';
 
     public get isStandalone(): boolean {
       return this.$route.meta.standalone;