+++ packages/components-js/src/utils/hasShadowDom.ts
@@ -1,3 +1,1 @@
-export function hasShadowDom(el: HTMLElement) {
-  return !!el.shadowRoot && !!(el as any).attachShadow;
-}
+export const hasShadowDom = (el: HTMLElement): boolean => !!el.shadowRoot && !!(el as any).attachShadow;