+++ packages/components-angular/projects/components-wrapper/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `Divider`
+
 ## [1.1.2] - 2020-03-17
 
 ### Changed