+++ packages/components-angular/projects/components-wrapper/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Improve form elements
+
 ## [1.2.0] - 2020-03-25
 
 ### Added