+++ packages/components-angular/projects/components-wrapper/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `Text List`
 
+### Added
+- New headline size `headline-5` to `p-headline` 
+
 ## [1.3.0-rc.0] - 2020-04-03
 
 ### Fixed