+++ packages/components-angular/projects/components-wrapper/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- New headline size `headline-5` to `p-headline` 
+
 ## [1.3.0-rc.0] - 2020-04-03
 
 ### Fixed