+++ packages/components-angular/projects/components-wrapper/CHANGELOG.md
@@ -6,10 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.3.0] - 2020-04-08
-
 ### Added
 - `Text List`
+
+## [1.3.0] - 2020-04-08
 
 ### Added
 - New headline size `headline-5` to `p-headline`